Extract goals lever input control (#921, #924)
